+++ sql-planner/src/ir/transformation/redistribution/tests.rs
@@ -323,7 +323,7 @@ fn test_slices_2() {
     insta::assert_snapshot!(plan.as_explain().unwrap(), @r#"
     projection (sum(("count_1"::int))::int -> "col_1")
         motion [policy: full, program: ReshardIfNeeded]
-            projection (count(("t2"."e"::int))::int -> "count_1")
+            projection (count(("t2"."e"::int::int))::int -> "count_1")
                 join on true::bool
                     scan "unnamed_subquery"
                         projection ("t2"."f"::int -> "f")
